requirements/dev.txt
-r base.txt
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
ruff
bandit
iil-testkit>=0.1
